refactor: splite CVPreparationSection.component and useCVPreparation.hook to smaller parts
