+++ client/src/components/Dashboard/LearnSection/Recruitment/components/CVPreparation/components/Common/index.ts
@@ -1,0 +1,7 @@
+export { CVErrorState } from './CVErrorState.component';
+export { CVMainView } from './CVMainView.component';
+export { CVDetailView } from './CVDetailView.component';
+export { CVStatsOverview } from './CVStatsOverview.component';
+export { CVInfoSection } from './CVInfoSection.component';
+export { CVHeader } from './CVHeader.component';
+export { CVSectionCard } from './CVSectionCard.component';
