+++ client/src/components/Dashboard/LearnSection/Recruitment/components/CVPreparation/hooks/index.ts
@@ -1,0 +1,4 @@
+export { useCVPreparation } from './useCVPreparation.hook';
+export { useCVData } from './useCVData.hook';
+export { useCVNavigation } from './useCVNavigation.hook';
+export { useCVProgress } from './useCVProgress.hook';
