+++ client/src/components/Dashboard/LearnSection/Recruitment/components/CVPreparation/utils/index.ts
@@ -1,0 +1,1 @@
+export { useCVSections, getCategoryMapping, getViewConfig } from './cvSections.utils';
